qna-chatbot/requirements.txt
langchain_openai
langchain
python-dotenv
langchain_community
langchain_core
streamlit